requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary  # для PostgreSQL
python-jose[cryptography]
passlib[bcrypt]
python-multipart
jinja2
alembic